+++ CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.12.0] - 2026-07-23
+
+### Added
+
+- Add exclusive end line and column positions to diagnostics when a source range is available
+
 ## [0.11.0] - 2026-04-07
 
 ### Breaking Changes
@@ -143,7 +149,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix path traversal in normalize output path
 - Fix godoc link to mib.DiagnosticConfig
 
-[Unreleased]: https://github.com/golangsnmp/gomib/compare/v0.11.0...HEAD
+[Unreleased]: https://github.com/golangsnmp/gomib/compare/v0.12.0...HEAD
+[0.12.0]: https://github.com/golangsnmp/gomib/compare/v0.11.0...v0.12.0
 [0.11.0]: https://github.com/golangsnmp/gomib/compare/v0.10.0...v0.11.0
 [0.10.0]: https://github.com/golangsnmp/gomib/compare/v0.9.0...v0.10.0
 [0.9.0]: https://github.com/golangsnmp/gomib/compare/v0.8.0...v0.9.0